requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pytest>=8.0
pybullet>=3.2.5
pyserial>=3.5
requests>=2.28.0
gunicorn>=21.0
python-dotenv>=1.0
twilio>=8.10.0
python-telegram-bot>=20.0

